requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-ckeditor-5==0.2.11
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
geographiclib==2.0
geopy==2.4.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
simplejson==3.19.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
